% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_HMM.R
\name{calculate_log_likelihood}
\alias{calculate_log_likelihood}
\title{Calculate log-likelihood for model fit}
\usage{
calculate_log_likelihood(data, model_fit)
}
\arguments{
\item{data}{numeric vector containing given data.}

\item{model_fit}{a list containing the model_fit. This model fit should contain the estimated mean \code{mu}, estimated variance \code{sigma}, and estimated transition matrix \code{matrix}.}
}
\value{
a double containing the log likelihood of the model for the data.
}
\description{
Gives the log-likelihood for a given model fit for multiple channels with a given data for gaussian emisions
}
