% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{ck_model}
\alias{ck_model}
\title{Transition matrix according to the Chung-Kennedy model}
\arguments{
\item{parameters}{vector with parameters of the ck model.}

\item{l}{integer containing the number of signals (channels).}
}
\value{
the transition matrix for a ck-model with the given parameters \eqn{\lambda}, \eqn{\eta}, \eqn{k}
}
\description{
Produces a transition matrix acording to the Chung-Kennedy (CK) model for a given vector of parameters.
}
\examples{
# For l=2
q<-ck_model(c(0.9,0.8,0.5), 2)
q
}
