% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_HMM.R
\name{HMM_custom}
\alias{HMM_custom}
\title{HMM model fit for Gaussian data with custom transition matrix, custom emission distribution, and associated losses.}
\usage{
HMM_custom(
  data,
  matrix_pars_estimation,
  matrix_from_pars,
  matrix_pars_init,
  emit_estimation,
  emit_generation,
  emit_init,
  max_it = 100,
  threshold = 1e-06
)
}
\arguments{
\item{data}{numeric vector containing given data.}

\item{matrix_pars_estimation}{function optimizing the parameters of a tranistion matrix model according to a custom loss function. The function should return 
a list of a vector \code{pars} with the optimized parameters, and a double \code{loss} with the minimized loss value.}

\item{matrix_from_pars}{function describing the transition matrix model given a vector of parameters. The function should return a squared transition matrix.}

\item{matrix_pars_init}{vector of parameters to initialize the \code{matrix_from_pars} function.}

\item{emit_estimation}{function optimizing the parameters of an emission model according to a custom loss function. The function should return a vector of
the optimized parameters.}

\item{emit_generation}{function calculating the probabilities of one data point to come from each hidden state, for a given distribution.
The function should return a vector of length equal to the sum state space with the probabilities of that data point to be in the corresponding state.}

\item{emit_init}{vector containing the initial parameters of the emission distribution. The two first components are the base level and step level size. The next \code{n_channels}+1 entries are the initial standard deviations for each level.}

\item{max_it}{integer containing the maximal number of iterations of the Baum-Welch-type algorithm. Set to defaul of 100.}

\item{threshold}{double containing the threshold of convergence for the Baum-Welch-type algorithm}
}
\value{
A list containing a vector \code{matrix_pars} of the estimated parameters for the given \code{model},
the estimated transition matrix \code{matrix}, a vector \code{mu} of the estimated levels, a vector \code{sigma} of the estimated standard deviations for each level,
a vector \code{pi} of the estimated initial distribution, and a double \code{loglik} of the pseudo-log-likelihood of this model.
}
\description{
Gives a model fit for multiple channels with a custom model of the transition matrix, custom emission distribution, and custom parameter optimization steps.
}
