% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_HMM.R
\name{HMM_estimate}
\alias{HMM_estimate}
\title{HMM model fit for Gaussian data}
\usage{
HMM_estimate(
  data,
  model,
  n_channels,
  emit_init,
  max_it = 100,
  threshold = 1e-06
)
}
\arguments{
\item{data}{numeric vector containing given data.}

\item{model}{string of model name to fit to the data. Currently supported "VND", Chung-Kennedy "CK", and uncoupled "independent"}

\item{n_channels}{integer of number of channels}

\item{emit_init}{vector containing the initial parameters of the emission distribution. The two first components are the base level and step level size. The next \code{n_channels}+1 entries are the initial standard deviations for each level.}

\item{max_it}{integer containing the maximal number of iterations of the Baum-Welch-type algorithm. Set to defaul of 100.}

\item{threshold}{double containing the threshold of convergence for the Baum-Welch-type algorithm}
}
\value{
A list containing a vector \code{matrix_pars} of the estimated parameters for the given \code{model},
the estimated transition matrix \code{matrix}, a vector \code{mu} of the estimated levels, a vector \code{sigma} of the estimated standard deviations for each level,
a vector \code{pi} of the estimated initial distribution, and a double \code{loglik} of the pseudo-log-likelihood of this model.
}
\description{
Gives a model fit for multiple channels with a given data and model for a HMM with Gaussian emissions and equidistant levels.
}
