% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{simulate_MC}
\alias{simulate_MC}
\title{Simulation of Markov chains}
\arguments{
\item{n_data}{integer containing the number of data points to be simulated.}

\item{P_trans}{transition matrix to be simulated from. It should be a square matrix.}

\item{intitial}{vector of length equal to the number of rows of \code{P_trans} containing the intital probability distribution.}
}
\value{
vector of length \code{n_data} containing the simulated markov chain with state space {0,1,...,\code{nrow(P_trans)}}.
}
\description{
Simulates \code{n_data} data points from a Markov chain with given transition matrix and intial probability.
}
\examples{
q <- vnd_model(c(0.9,0.8,0.98,0.89))
mc <- simulate_MC(100, q, c(1/3,1/3,1/3))
plot(mc, type = "s")
}
