% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{uc_model}
\alias{uc_model}
\title{Transition matrix according to the uncoupled model}
\arguments{
\item{parameters}{vector with parameters of the uc model.}

\item{l}{integer containing the number of signals (channels).}
}
\value{
the transition matrix for a uc-model with the given parameters \eqn{\lambda}, \eqn{\eta}
}
\description{
Produces a transition matrix acording to the uncoupled (independent) model for a given vector of parameters.
}
\examples{
# For l=2
q <- uc_model(c(0.9,0.8), 2)
q
}
