% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{Viterbi_simple}
\alias{Viterbi_simple}
\title{Viterbi path}
\arguments{
\item{y}{numeric vector of data}

\item{K}{integer containing the number of states}

\item{pi}{vector containing the initial probability of the system}

\item{P_trans}{KxK matrix of tranisitiong probabilities}

\item{mu}{vector containing the level for each state}

\item{var}{vector containing the variance for each state}
}
\value{
vector of length equal to the length of the data containing the viterbi path
}
\description{
Produces a the estimated viterbi path for given data and parameters.
}
