% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{vnd_model}
\alias{vnd_model}
\title{Transition matrix according to the VND model}
\arguments{
\item{parameters}{vector with parameters of the VND model \eqn{(\lambda_0,\eta_1,\lamda_1,\ldtos,\eta_\ell)}.}
}
\value{
the transition matrix for a VND model with the given parameters
}
\description{
Produces a transition matrix acording to the VND model for a given vector of parameters.
}
\examples{
# For l=2
q<-vnd_model(c(0.9,0.8,0.98,0.89))
q
}
